applying jarvis patch

+++ test.c
@@ -1,8 +1,8 @@
 
 int test(int weight) {
     int a = 1;
-    float b = 2.0;
-    return a + b + weight;
+    float b = 2.0f; // Ensure the float is specified with 'f'
+    return a + static_cast<int>(b) + weight; // Cast 'b' to 'int'
 }
 
 int main(int argc, char *argv[]) {
